moved testingWithGoogle to tests, factored out some test helpers into Equationcomponent_test_helpers
