+++ CMakeLists.txt
@@ -126,7 +126,7 @@ else()
 endif()
 
 add_subdirectory(src)
-add_subdirectory(testingWithGoogle)
+add_subdirectory(tests)
