+++ tests/AuxUnitConversionTest.cpp
