+++ tests/PowerTest.cpp
